Convert X Rebirth extension to the Game Description Language
Replace the hand-written X Rebirth game extension with a declarative
game.yaml compiled by the GDL toolchain (added as the
game-description-language submodule). The extension is now a game.yaml,
game art, and a small src/hooks.ts holding the two imperative pieces GDL
can't express declaratively: the content.xml installer (parses XML and
emits attribute instructions) and the three in-game health checks.

This is a 1:1 behavioral replacement. The game-extension-test harness
produces identical results against every X Rebirth mod on Nexus (956
passed, 15 skipped) for both the original and converted extension.

Also teaches @vortex/game-extension-test to load GDL extensions (detects
game.yaml, loads the generated entry through the harness module aliases,
and captures health checks registered via registerHealthCheck), so future
GDL game conversions are covered by the same regression harness.

The GDL submodule tracks feat/vortex-xrebirth-integration, which adds the
features this conversion needs (declarative installer predicates, a custom
install hook, runtime diagnostics, stop patterns, and Steam app id
derivation).
