+++ packages/game-extension-test/vitest.fixtures.config.ts
@@ -1,4 +1,6 @@
 import { createRequire } from "node:module";
+import * as path from "node:path";
+import { fileURLToPath } from "node:url";
 
 import { defineConfig } from "vitest/config";
 
@@ -8,6 +10,12 @@ import { defineConfig } from "vitest/config";
  */
 const require_ = createRequire(import.meta.url);
 const VORTEX_API_MOCK = require_.resolve("@vortex/extension-test-mocks");
+
+// GDL-generated extensions import the runtime via the `@gdl/runtime` webpack
+// alias. The harness loads the generated TS (not the bundle), so map the same
+// specifier to the submodule's runtime source.
+const here = path.dirname(fileURLToPath(import.meta.url));
+const GDL_RUNTIME = path.resolve(here, "../../game-description-language/src/runtime/index.ts");
 
 // Each fixture makes at least two HTTP calls (listModFiles + manifest); a slow
 // Nexus response can easily exceed vitest's 5s default.
@@ -19,7 +27,10 @@ const FIXTURE_MAX_CONCURRENCY = 24;
 
 export default defineConfig({
   resolve: {
-    alias: [{ find: /^@nexusmods\/vortex-api$/, replacement: VORTEX_API_MOCK }],
+    alias: [
+      { find: /^@nexusmods\/vortex-api$/, replacement: VORTEX_API_MOCK },
+      { find: /^@gdl\/runtime$/, replacement: GDL_RUNTIME },
+    ],
   },
   test: {
     environment: "node",
